Чинить импорт BOQ: понятные ошибки в модалке вместо ложного «успешно» + фикс краша
Фронт (оба пути импорта — single-position и mass): убрать зелёный «успешно»,
который показывался после любой ошибки (вывод из uploadProgress), на явный
importStatus/importError; при ошибке — красный Alert с причиной (Строка N: …),
модалка не закрывается, mass при частичной загрузке не отдаёт «успех». Общая
пред-валидация строк (src/utils/boq/importRowValidation.ts): кол-во ≤ 0 и расход
< 1 — блокирующая ошибка, пустые коэффициенты — предупреждение. Обновлять таблицу
при закрытии независимо от успеха (иначе закоммиченные строки не видны).

Бэкенд: фикс краша импорта (COALESCE($15, 0::numeric) — голый 0 пиннил параметр
в int4, total_amount переполнял); маппинг FK единицы измерения в 4xx для
Create/Update Work/Material Name (не 500 + Sentry).

+++ backend/internal/repository/import_boq.go
@@ -259,7 +259,7 @@ func (r *ImportRepo) BulkImport(ctx context.Context, in ImportInput) (*ImportRes
 			$12,
 			$13,
 			COALESCE($14::public.currency_type, 'RUB'::public.currency_type),
-			COALESCE($15, 0),
+			COALESCE($15, 0::numeric),
 			$16::public.delivery_price_type,
 			$17,
 			$18,
